shift logic int service
